catalog-backend: split up catalog database into 3 separate implementations
