+++ plugins/catalog-backend/src/database/DefaultProcessingDatabase.ts
@@ -15,7 +15,7 @@
  */
 
 import { Entity, stringifyEntityRef } from '@backstage/catalog-model';
-import { ConflictError, NotFoundError } from '@backstage/errors';
+import { ConflictError } from '@backstage/errors';
 import { Knex } from 'knex';
 import lodash from 'lodash';
 import { v4 as uuid } from 'uuid';
@@ -25,15 +25,10 @@ import {
   GetProcessableEntitiesResult,
   ProcessingDatabase,
   RefreshStateItem,
-  RefreshOptions,
-  ReplaceUnprocessedEntitiesOptions,
   UpdateProcessedEntityOptions,
-  ListAncestorsOptions,
-  ListAncestorsResult,
   UpdateEntityCacheOptions,
   ListParentsOptions,
   ListParentsResult,
-  RefreshByKeyOptions,
 } from './types';
 import { ProcessingIntervalFunction } from '../processing/refresh';
 import { rethrowError, timestampToDateTime } from './conversion';
@@ -54,7 +49,6 @@ import { DeferredEntity } from '@backstage/plugin-catalog-node';
 // errors in the underlying engine due to exceeding query limits, but large
 // enough to get the speed benefits.
 const BATCH_SIZE = 50;
-const MAX_ANCESTOR_DEPTH = 32;
 
 export class DefaultProcessingDatabase implements ProcessingDatabase {
   constructor(
@@ -193,236 +187,6 @@ export class DefaultProcessingDatabase implements ProcessingDatabase {
       .where('entity_id', id);
   }
 
-  async replaceUnprocessedEntities(
-    txOpaque: Transaction,
-    options: ReplaceUnprocessedEntitiesOptions,
-  ): Promise<void> {
-    const tx = txOpaque as Knex.Transaction;
-
-    const { toAdd, toUpsert, toRemove } = await this.createDelta(tx, options);
-
-    if (toRemove.length) {
-      let removedCount = 0;
-      const rootId = () => {
-        if (tx.client.config.client.includes('mysql')) {
-          return tx.raw('CAST(NULL as UNSIGNED INT)', []);
-        }
-
-        return tx.raw('CAST(NULL as INT)', []);
-      };
-      for (const refs of lodash.chunk(toRemove, 1000)) {
-        /*
-      WITH RECURSIVE
-        -- All the nodes that can be reached downwards from our root
-        descendants(root_id, entity_ref) AS (
-          SELECT id, target_entity_ref
-          FROM refresh_state_references
-          WHERE source_key = "R1" AND target_entity_ref = "A"
-          UNION
-          SELECT descendants.root_id, target_entity_ref
-          FROM descendants
-          JOIN refresh_state_references ON source_entity_ref = descendants.entity_ref
-        ),
-        -- All the nodes that can be reached upwards from the descendants
-        ancestors(root_id, via_entity_ref, to_entity_ref) AS (
-          SELECT CAST(NULL as INT), entity_ref, entity_ref
-          FROM descendants
-          UNION
-          SELECT
-            CASE WHEN source_key IS NOT NULL THEN id ELSE NULL END,
-            source_entity_ref,
-            ancestors.to_entity_ref
-          FROM ancestors
-          JOIN refresh_state_references ON target_entity_ref = ancestors.via_entity_ref
-        )
-      -- Start out with all of the descendants
-      SELECT descendants.entity_ref
-      FROM descendants
-      -- Expand with all ancestors that point to those, but aren't the current root
-      LEFT OUTER JOIN ancestors
-        ON ancestors.to_entity_ref = descendants.entity_ref
-        AND ancestors.root_id IS NOT NULL
-        AND ancestors.root_id != descendants.root_id
-      -- Exclude all lines that had such a foreign ancestor
-      WHERE ancestors.root_id IS NULL;
-      */
-        removedCount += await tx<DbRefreshStateRow>('refresh_state')
-          .whereIn('entity_ref', function orphanedEntityRefs(orphans) {
-            return (
-              orphans
-                // All the nodes that can be reached downwards from our root
-                .withRecursive('descendants', function descendants(outer) {
-                  return outer
-                    .select({ root_id: 'id', entity_ref: 'target_entity_ref' })
-                    .from('refresh_state_references')
-                    .where('source_key', options.sourceKey)
-                    .whereIn('target_entity_ref', refs)
-                    .union(function recursive(inner) {
-                      return inner
-                        .select({
-                          root_id: 'descendants.root_id',
-                          entity_ref:
-                            'refresh_state_references.target_entity_ref',
-                        })
-                        .from('descendants')
-                        .join('refresh_state_references', {
-                          'descendants.entity_ref':
-                            'refresh_state_references.source_entity_ref',
-                        });
-                    });
-                })
-                // All the nodes that can be reached upwards from the descendants
-                .withRecursive('ancestors', function ancestors(outer) {
-                  return outer
-                    .select({
-                      root_id: rootId(),
-                      via_entity_ref: 'entity_ref',
-                      to_entity_ref: 'entity_ref',
-                    })
-                    .from('descendants')
-                    .union(function recursive(inner) {
-                      return inner
-                        .select({
-                          root_id: tx.raw(
-                            'CASE WHEN source_key IS NOT NULL THEN id ELSE NULL END',
-                            [],
-                          ),
-                          via_entity_ref: 'source_entity_ref',
-                          to_entity_ref: 'ancestors.to_entity_ref',
-                        })
-                        .from('ancestors')
-                        .join('refresh_state_references', {
-                          target_entity_ref: 'ancestors.via_entity_ref',
-                        });
-                    });
-                })
-                // Start out with all of the descendants
-                .select('descendants.entity_ref')
-                .from('descendants')
-                // Expand with all ancestors that point to those, but aren't the current root
-                .leftOuterJoin('ancestors', function keepaliveRoots() {
-                  this.on(
-                    'ancestors.to_entity_ref',
-                    '=',
-                    'descendants.entity_ref',
-                  );
-                  this.andOnNotNull('ancestors.root_id');
-                  this.andOn('ancestors.root_id', '!=', 'descendants.root_id');
-                })
-                .whereNull('ancestors.root_id')
-            );
-          })
-          .delete();
-
-        await tx<DbRefreshStateReferencesRow>('refresh_state_references')
-          .where('source_key', '=', options.sourceKey)
-          .whereIn('target_entity_ref', refs)
-          .delete();
-      }
-
-      this.options.logger.debug(
-        `removed, ${removedCount} entities: ${JSON.stringify(toRemove)}`,
-      );
-    }
-
-    if (toAdd.length) {
-      // The reason for this chunking, rather than just massively batch
-      // inserting the entire payload, is that we fall back to the individual
-      // upsert mechanism below on conflicts. That path is massively slower than
-      // the fast batch path, so we don't want to end up accidentally having to
-      // for example item-by-item upsert tens of thousands of entities in a
-      // large initial delivery dump. The implication is that the size of these
-      // chunks needs to weigh the benefit of fast successful inserts, against
-      // the drawback of super slow but more rare fallbacks. There's quickly
-      // diminishing returns though with turning up this value way high.
-      for (const chunk of lodash.chunk(toAdd, 50)) {
-        try {
-          await tx.batchInsert(
-            'refresh_state',
-            chunk.map(item => ({
-              entity_id: uuid(),
-              entity_ref: stringifyEntityRef(item.deferred.entity),
-              unprocessed_entity: JSON.stringify(item.deferred.entity),
-              unprocessed_hash: item.hash,
-              errors: '',
-              location_key: item.deferred.locationKey,
-              next_update_at: tx.fn.now(),
-              last_discovery_at: tx.fn.now(),
-            })),
-            BATCH_SIZE,
-          );
-          await tx.batchInsert(
-            'refresh_state_references',
-            chunk.map(item => ({
-              source_key: options.sourceKey,
-              target_entity_ref: stringifyEntityRef(item.deferred.entity),
-            })),
-            BATCH_SIZE,
-          );
-        } catch (error) {
-          if (!isDatabaseConflictError(error)) {
-            throw error;
-          } else {
-            this.options.logger.debug(
-              `Fast insert path failed, falling back to slow path, ${error}`,
-            );
-            toUpsert.push(...chunk);
-          }
-        }
-      }
-    }
-
-    if (toUpsert.length) {
-      for (const {
-        deferred: { entity, locationKey },
-        hash,
-      } of toUpsert) {
-        const entityRef = stringifyEntityRef(entity);
-
-        try {
-          let ok = await this.updateUnprocessedEntity(
-            tx,
-            entity,
-            hash,
-            locationKey,
-          );
-          if (!ok) {
-            ok = await this.insertUnprocessedEntity(
-              tx,
-              entity,
-              hash,
-              locationKey,
-            );
-          }
-
-          if (ok) {
-            await tx<DbRefreshStateReferencesRow>(
-              'refresh_state_references',
-            ).insert({
-              source_key: options.sourceKey,
-              target_entity_ref: entityRef,
-            });
-          } else {
-            const conflictingKey = await this.checkLocationKeyConflict(
-              tx,
-              entityRef,
-              locationKey,
-            );
-            if (conflictingKey) {
-              this.options.logger.warn(
-                `Source ${options.sourceKey} detected conflicting entityRef ${entityRef} already referenced by ${conflictingKey} and now also ${locationKey}`,
-              );
-            }
-          }
-        } catch (error) {
-          this.options.logger.error(
-            `Failed to add '${entityRef}' from source '${options.sourceKey}', ${error}`,
-          );
-        }
-      }
-    }
-  }
-
   async getProcessableEntities(
     txOpaque: Transaction,
     request: { processBatchSize: number },
@@ -487,45 +251,6 @@ export class DefaultProcessingDatabase implements ProcessingDatabase {
     };
   }
 
-  async listAncestors(
-    txOpaque: Transaction,
-    options: ListAncestorsOptions,
-  ): Promise<ListAncestorsResult> {
-    const tx = txOpaque as Knex.Transaction;
-    const { entityRef } = options;
-    const entityRefs = new Array<string>();
-
-    let currentRef = entityRef.toLocaleLowerCase('en-US');
-    for (let depth = 1; depth <= MAX_ANCESTOR_DEPTH; depth += 1) {
-      const rows = await tx<DbRefreshStateReferencesRow>(
-        'refresh_state_references',
-      )
-        .where({ target_entity_ref: currentRef })
-        .select();
-
-      if (rows.length === 0) {
-        if (depth === 1) {
-          throw new NotFoundError(`Entity ${currentRef} not found`);
-        }
-        throw new NotFoundError(
-          `Entity ${entityRef} has a broken parent reference chain at ${currentRef}`,
-        );
-      }
-
-      const parentRef = rows.find(r => r.source_entity_ref)?.source_entity_ref;
-      if (!parentRef) {
-        // We've reached the top of the tree which is the entityProvider.
-        // In this case we refresh the entity itself.
-        return { entityRefs };
-      }
-      entityRefs.push(parentRef);
-      currentRef = parentRef;
-    }
-    throw new Error(
-      `Unable receive ancestors for ${entityRef}, reached maximum depth of ${MAX_ANCESTOR_DEPTH}`,
-    );
-  }
-
   async listParents(
     txOpaque: Transaction,
     options: ListParentsOptions,
@@ -541,37 +266,6 @@ export class DefaultProcessingDatabase implements ProcessingDatabase {
     const entityRefs = rows.map(r => r.source_entity_ref!).filter(Boolean);
 
     return { entityRefs };
-  }
-
-  async refresh(txOpaque: Transaction, options: RefreshOptions): Promise<void> {
-    const tx = txOpaque as Knex.Transaction;
-    const { entityRef } = options;
-
-    const updateResult = await tx<DbRefreshStateRow>('refresh_state')
-      .where({ entity_ref: entityRef.toLocaleLowerCase('en-US') })
-      .update({ next_update_at: tx.fn.now() });
-    if (updateResult === 0) {
-      throw new NotFoundError(`Failed to schedule ${entityRef} for refresh`);
-    }
-  }
-
-  async refreshByRefreshKeys(
-    txOpaque: Transaction,
-    options: RefreshByKeyOptions,
-  ) {
-    const tx = txOpaque as Knex.Transaction;
-    const { keys } = options;
-
-    await tx<DbRefreshStateRow>('refresh_state')
-      .whereIn('entity_id', function selectEntityRefs(tx2) {
-        tx2
-          .whereIn('key', keys)
-          .select({
-            entity_id: 'refresh_keys.entity_id',
-          })
-          .from('refresh_keys');
-      })
-      .update({ next_update_at: tx.fn.now() });
   }
 
   async transaction<T>(fn: (tx: Transaction) => Promise<T>): Promise<T> {
@@ -719,84 +413,6 @@ export class DefaultProcessingDatabase implements ProcessingDatabase {
       rows,
       r => `${r.source_entity_ref}:${r.target_entity_ref}:${r.type}`,
     );
-  }
-
-  private async createDelta(
-    tx: Knex.Transaction,
-    options: ReplaceUnprocessedEntitiesOptions,
-  ): Promise<{
-    toAdd: { deferred: DeferredEntity; hash: string }[];
-    toUpsert: { deferred: DeferredEntity; hash: string }[];
-    toRemove: string[];
-  }> {
-    if (options.type === 'delta') {
-      return {
-        toAdd: [],
-        toUpsert: options.added.map(e => ({
-          deferred: e,
-          hash: generateStableHash(e.entity),
-        })),
-        toRemove: options.removed.map(e => e.entityRef),
-      };
-    }
-
-    // Grab all of the existing references from the same source, and their locationKeys as well
-    const oldRefs = await tx<DbRefreshStateReferencesRow>(
-      'refresh_state_references',
-    )
-      .leftJoin<DbRefreshStateRow>('refresh_state', {
-        target_entity_ref: 'entity_ref',
-      })
-      .where({ source_key: options.sourceKey })
-      .select({
-        target_entity_ref: 'refresh_state_references.target_entity_ref',
-        location_key: 'refresh_state.location_key',
-        unprocessed_hash: 'refresh_state.unprocessed_hash',
-      });
-
-    const items = options.items.map(deferred => ({
-      deferred,
-      ref: stringifyEntityRef(deferred.entity),
-      hash: generateStableHash(deferred.entity),
-    }));
-
-    const oldRefsSet = new Map(
-      oldRefs.map(r => [
-        r.target_entity_ref,
-        {
-          locationKey: r.location_key,
-          oldEntityHash: r.unprocessed_hash,
-        },
-      ]),
-    );
-    const newRefsSet = new Set(items.map(item => item.ref));
-
-    const toAdd = new Array<{ deferred: DeferredEntity; hash: string }>();
-    const toUpsert = new Array<{ deferred: DeferredEntity; hash: string }>();
-    const toRemove = oldRefs
-      .map(row => row.target_entity_ref)
-      .filter(ref => !newRefsSet.has(ref));
-
-    for (const item of items) {
-      const oldRef = oldRefsSet.get(item.ref);
-      const upsertItem = { deferred: item.deferred, hash: item.hash };
-      if (!oldRef) {
-        // Add any entity that does not exist in the database
-        toAdd.push(upsertItem);
-      } else if (
-        (oldRef?.locationKey ?? undefined) !==
-        (item.deferred.locationKey ?? undefined)
-      ) {
-        // Remove and then re-add any entity that exists, but with a different location key
-        toRemove.push(item.ref);
-        toAdd.push(upsertItem);
-      } else if (oldRef.oldEntityHash !== item.hash) {
-        // Entities with modifications should be pushed through too
-        toUpsert.push(upsertItem);
-      }
-    }
-
-    return { toAdd, toUpsert, toRemove };
   }
 
   /**
