+++ plugins/catalog-backend/src/database/types.ts
@@ -109,16 +109,11 @@ export type ListParentsResult = {
   entityRefs: string[];
 };
 
+/**
+ * The database abstraction layer for Entity Processor interactions.
+ */
 export interface ProcessingDatabase {
   transaction<T>(fn: (tx: Transaction) => Promise<T>): Promise<T>;
-
-  /**
-   * Add unprocessed entities to the front of the processing queue using a mutation.
-   */
-  replaceUnprocessedEntities(
-    txOpaque: Transaction,
-    options: ReplaceUnprocessedEntitiesOptions,
-  ): Promise<void>;
 
   getProcessableEntities(
     txOpaque: Transaction,
@@ -152,10 +147,25 @@ export interface ProcessingDatabase {
     options: UpdateProcessedEntityErrorsOptions,
   ): Promise<void>;
 
+  listParents(
+    txOpaque: Transaction,
+    options: ListParentsOptions,
+  ): Promise<ListParentsResult>;
+}
+
+/**
+ * The database abstraction layer for Entity Provider interactions.
+ */
+export interface ProviderDatabase {
+  transaction<T>(fn: (tx: Transaction) => Promise<T>): Promise<T>;
+
   /**
-   * Schedules a refresh of a given entityRef.
+   * Add unprocessed entities to the front of the processing queue using a mutation.
    */
-  refresh(txOpaque: Transaction, options: RefreshOptions): Promise<void>;
+  replaceUnprocessedEntities(
+    txOpaque: Transaction,
+    options: ReplaceUnprocessedEntitiesOptions,
+  ): Promise<void>;
 
   /**
    * Schedules a refresh for every entity that has a matching set of refresh key stored for it.
@@ -164,6 +174,14 @@ export interface ProcessingDatabase {
     txOpaque: Transaction,
     options: RefreshByKeyOptions,
   ): Promise<void>;
+}
+
+// TODO(Rugvip): This is only partial for now
+/**
+ * The database abstraction layer for catalog access.
+ */
+export interface CatalogDatabase {
+  transaction<T>(fn: (tx: Transaction) => Promise<T>): Promise<T>;
 
   /**
    * Lists all ancestors of a given entityRef.
@@ -175,8 +193,8 @@ export interface ProcessingDatabase {
     options: ListAncestorsOptions,
   ): Promise<ListAncestorsResult>;
 
-  listParents(
-    txOpaque: Transaction,
-    options: ListParentsOptions,
-  ): Promise<ListParentsResult>;
+  /**
+   * Schedules a refresh of a given entityRef.
+   */
+  refresh(txOpaque: Transaction, options: RefreshOptions): Promise<void>;
 }
