+++ plugins/catalog-backend/src/service/AuthorizedEntitiesCatalog.test.ts
@@ -28,6 +28,8 @@ describe('AuthorizedEntitiesCatalog', () => {
     removeEntityByUid: jest.fn(),
     entityAncestry: jest.fn(),
     facets: jest.fn(),
+    refresh: jest.fn(),
+    listAncestors: jest.fn(),
   };
   const fakePermissionApi = {
     authorize: jest.fn(),
